+++ packages/w3wallets/src/metamask/types.ts
@@ -1,0 +1,6 @@
+export type NetworkSettings = {
+  name: string;
+  rpc: string;
+  chainId: number;
+  currencySymbol: string;
+};