+++ packages/w3wallets/tests/fixtures/base.ts
@@ -1,7 +1,11 @@
 import { withWallets } from "w3wallets";
 import { test as base } from "@playwright/test";
 
-export const test = withWallets(base, 'backpack', 'polkadotJS').extend<BaseFixture>({
+export const test = withWallets(
+  base,
+  "backpack",
+  "polkadotJS",
+).extend<BaseFixture>({
   magic: (_, use) => use(42),
 });