+++ packages/w3wallets/CHANGELOG.md
@@ -1,5 +1,17 @@
 # w3wallets
 
+## 0.4.1
+
+### Patch Changes
+
+- Fix metamask extension path
+
+## 0.4.0
+
+### Minor Changes
+
+- Add basic Metamask support
+
 ## 0.3.0
 
 ### Minor Changes